Revert "Move GeoJsonType enum into its own file"
This reverts commit f571a744aff6a3e19e34a61aa3f115e43313e37d.
