+++ src/main/java/org/openstreetmap/atlas/geography/geojson/GeoJsonBuilder.java
@@ -29,6 +29,44 @@ import com.google.gson.JsonPrimitive;
 public class GeoJsonBuilder
 {
     /**
+     * @author matthieun
+     * @author mgostintsev
+     */
+    public enum GeoJsonType
+    {
+        POINT("Point"),
+        LINESTRING("LineString"),
+        POLYGON("Polygon"),
+        MULTI_POINT("MultiPoint"),
+        MULTI_LINESTRING("MultiLineString"),
+        MULTI_POLYGON("MultiPolygon");
+
+        private final String type;
+
+        public static GeoJsonType forType(final String type)
+        {
+            for (final GeoJsonType value : values())
+            {
+                if (value.getType().equals(type))
+                {
+                    return value;
+                }
+            }
+            throw new CoreException("Invalid geoJson type: {}", type);
+        }
+
+        GeoJsonType(final String type)
+        {
+            this.type = type;
+        }
+
+        public String getType()
+        {
+            return this.type;
+        }
+    }
+
+    /**
      * Java bean to store geometry (as an {@link Iterable} of {@link Location}s) and properties as a
      * {@link String} to {@link Object} {@link Map}
      *
