+++ src/test/java/org/openstreetmap/atlas/geography/geojson/GeoJsonBuilderTest.java
@@ -12,6 +12,7 @@ import org.junit.Test;
 import org.openstreetmap.atlas.geography.Location;
 import org.openstreetmap.atlas.geography.PolyLine;
 import org.openstreetmap.atlas.geography.Polygon;
+import org.openstreetmap.atlas.geography.geojson.GeoJsonBuilder.GeoJsonType;
 import org.openstreetmap.atlas.geography.geojson.GeoJsonBuilder.GeometryWithProperties;
 import org.openstreetmap.atlas.geography.geojson.GeoJsonBuilder.LocationIterableProperties;
 import org.openstreetmap.atlas.streaming.readers.GeoJsonReader;
